Move providers to /src/components/providers directory
